requirements.txt
Django==1.3.1
South==0.7.3
django-cronjobs
requests==0.6.6
django-registration==0.7
twilio==3.3.2
